FilterBar refactor to add controlled from outside option
